[#49]: Reorganize Record tab into a record module
Split the monolithic Record tab into components, hooks, and utils under
drafting/record and colocate useVerseRecorder with its screen.

Refs #49
