+++ src/app/tabs/drafting/record/hooks/useRecordTabGuards.test.ts
@@ -1,0 +1,146 @@
+import { Alert } from 'react-native';
+import { act, renderHook } from '@testing-library/react-native';
+import { useRecordTabGuards } from './useRecordTabGuards';
+
+const mockNavigation = {
+  addListener: jest.fn(() => jest.fn()),
+  dispatch: jest.fn(),
+};
+
+describe('useRecordTabGuards', () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
+
+  it('withPausedGuard shows an alert instead of running the action while paused', () => {
+    const alertSpy = jest.spyOn(Alert, 'alert').mockImplementation(() => {});
+    const action = jest.fn();
+
+    const { result } = renderHook(() =>
+      useRecordTabGuards({
+        status: 'paused',
+        permission: 'granted',
+        requestPermission: jest.fn(),
+        discardPaused: jest.fn().mockResolvedValue(undefined),
+        navigation: mockNavigation as never,
+      }),
+    );
+
+    result.current.withPausedGuard(action);
+
+    expect(alertSpy).toHaveBeenCalledWith(
+      'Recording in progress',
+      expect.stringContaining('switching verses'),
+      expect.any(Array),
+    );
+    expect(action).not.toHaveBeenCalled();
+
+    alertSpy.mockRestore();
+  });
+
+  it('withPausedGuard runs the action immediately when not paused', () => {
+    const action = jest.fn();
+
+    const { result } = renderHook(() =>
+      useRecordTabGuards({
+        status: 'idle',
+        permission: 'granted',
+        requestPermission: jest.fn(),
+        discardPaused: jest.fn().mockResolvedValue(undefined),
+        navigation: mockNavigation as never,
+      }),
+    );
+
+    result.current.withPausedGuard(action);
+
+    expect(action).toHaveBeenCalledTimes(1);
+  });
+
+  it('withPausedGuard discards the paused take before running the action', async () => {
+    const alertSpy = jest.spyOn(Alert, 'alert').mockImplementation(() => {});
+    const discardPaused = jest.fn().mockResolvedValue(undefined);
+    const action = jest.fn();
+
+    const { result } = renderHook(() =>
+      useRecordTabGuards({
+        status: 'paused',
+        permission: 'granted',
+        requestPermission: jest.fn(),
+        discardPaused,
+        navigation: mockNavigation as never,
+      }),
+    );
+
+    result.current.withPausedGuard(action);
+
+    const buttons = alertSpy.mock.calls[0]![2] as Array<{
+      text: string;
+      onPress?: () => Promise<void>;
+    }>;
+    const discardButton = buttons.find(button => button.text === 'Discard');
+    await act(async () => {
+      await discardButton?.onPress?.();
+    });
+
+    expect(discardPaused).toHaveBeenCalledTimes(1);
+    expect(action).toHaveBeenCalledTimes(1);
+
+    alertSpy.mockRestore();
+  });
+
+  it('registers a beforeRemove listener while paused', () => {
+    renderHook(() =>
+      useRecordTabGuards({
+        status: 'paused',
+        permission: 'granted',
+        requestPermission: jest.fn(),
+        discardPaused: jest.fn().mockResolvedValue(undefined),
+        navigation: mockNavigation as never,
+      }),
+    );
+
+    expect(mockNavigation.addListener).toHaveBeenCalledWith(
+      'beforeRemove',
+      expect.any(Function),
+    );
+  });
+
+  it('ensureMicPermission returns true when permission is granted', async () => {
+    const { result } = renderHook(() =>
+      useRecordTabGuards({
+        status: 'idle',
+        permission: 'granted',
+        requestPermission: jest.fn(),
+        discardPaused: jest.fn().mockResolvedValue(undefined),
+        navigation: mockNavigation as never,
+      }),
+    );
+
+    await expect(result.current.ensureMicPermission()).resolves.toBe(true);
+  });
+
+  it('ensureMicPermission shows Settings alert when permission is blocked', async () => {
+    const alertSpy = jest.spyOn(Alert, 'alert').mockImplementation(() => {});
+    const requestPermission = jest.fn();
+
+    const { result } = renderHook(() =>
+      useRecordTabGuards({
+        status: 'idle',
+        permission: 'blocked',
+        requestPermission,
+        discardPaused: jest.fn().mockResolvedValue(undefined),
+        navigation: mockNavigation as never,
+      }),
+    );
+
+    await expect(result.current.ensureMicPermission()).resolves.toBe(false);
+    expect(alertSpy).toHaveBeenCalledWith(
+      'Microphone access required',
+      expect.stringContaining('Settings'),
+      expect.any(Array),
+    );
+    expect(requestPermission).not.toHaveBeenCalled();
+
+    alertSpy.mockRestore();
+  });
+});
